back/requirements.txt
flask_restplus
google-cloud-language
google-cloud-speech
spacy
pyphen
pandas
Pillow
